+++ cryptofeed/exceptions.py
@@ -49,5 +49,6 @@ class RestResponseError(Exception):
 class ConnectionClosed(Exception):
     pass
 
+
 class ConnectionExists(Exception):
     pass